refactor(storage): Split cost.rs into smaller modules (#4321)
Split the 2046-line cost.rs into logical modules for better maintainability:
- cost/estimator.rs (140 lines): CostEstimator struct and parameters
- cost/types.rs (192 lines): AccessMethod, TableIndexInfo, row size estimation
- cost/scan.rs (362 lines): Table scan, index scan, skip-scan cost estimation
- cost/dml.rs (228 lines): INSERT, UPDATE, DELETE cost estimation
- cost/mod.rs (36 lines): Module re-exports
- cost/tests.rs (1142 lines): All existing tests

No public API changes. All 45 existing tests pass.

Closes #4321
